- 펌웨어를 한개 화일로 묶게 하였음(2단계). 오유가 있어 펌웨어가 묶어지지 않는 경우 한개 화일로 묶지 않게 하였음. images.zip화일의 유무판단을 진행하여 콤파일오유를 판단하게 하였음.

+++ pack/main.cpp
@@ -170,9 +170,22 @@ int pack_firm()
     const char *pack_file_names[] = {"boot","config.yaml","fip_fsbl.bin","imtb","prim","pst","pusr1","pusr2","pwx"};
     char str_afile_name[256];
     //decompress images.zip
+    system("rm -f ./images.img");
+    system("rm -f ./images.zip");
     system("rm -rf ./images");
     system("mkdir ./images");
+    system("unzip ./images.zip -d ./images");
     system("cp -f ../solutions/usb_cam/generated/images.zip ./");
+
+    FILE* tf = NULL;
+    tf = fopen("./images.zip", "rb");
+    if (tf == NULL)
+    {
+        printf ("ERROR! file not found images.zip\n");
+        return 0;
+    }
+    fclose(tf);
+
     system("unzip ./images.zip -d ./images");
     system("rm -f ./images/misc");
     system("rm -f ./images/boot0");
